Fix #2754

+++ config/cache.php
@@ -84,7 +84,7 @@ return [
             'servers'       => [
                 [
                     'host'   => env('MEMCACHED_HOST', '127.0.0.1'),
-                    'port'   => env('MEMCACHED_PORT', 11211),
+                    'port'   => (int)env('MEMCACHED_PORT', 11211),
                     'weight' => 100,
                 ],
             ],
